float32s

+++ feudal_trainer.py
@@ -159,10 +159,10 @@ parser.add_argument(
 if __name__== "__main__":
     args = parser.parse_args()
 
-    upper_level_obs_space = spaces.Box(low = -np.inf, high = np.inf, shape = (20,), dtype = np.float64)
-    upper_level_action_space = spaces.Box(low = -1, high = 1, shape = (5,), dtype = np.float64)
-    lower_level_obs_space = spaces.Box(low=-np.inf, high=np.inf, shape=(3,), dtype=np.float64)
-    lower_level_action_space = spaces.Box(low = -1, high = 1, shape = (2,), dtype = np.float64)
+    upper_level_obs_space = spaces.Box(low = -np.inf, high = np.inf, shape = (20,), dtype = np.float32)
+    upper_level_action_space = spaces.Box(low = -1, high = 1, shape = (5,), dtype = np.float32)
+    lower_level_obs_space = spaces.Box(low=-np.inf, high=np.inf, shape=(3,), dtype=np.float32)
+    lower_level_action_space = spaces.Box(low = -1, high = 1, shape = (2,), dtype = np.float32)
 
     policies = {"upper_level_agent": (
         None, upper_level_obs_space,
